% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algo_leverage.R
\name{algo_leverage}
\alias{algo_leverage}
\title{Algorithmic Leveraging Function
implements algorithmic leveraging for linear regression using uniform and leverage score based on subsampling of rows}
\usage{
algo_leverage(xi, yi, r = floor(0.2 * length(yi)))
}
\arguments{
\item{xi}{independent variables}

\item{yi}{response variable}

\item{r}{the sampling size,default=20% of the input data}
}
\value{
b_est the estimated regression coefficients
}
\description{
Algorithmic Leveraging Function
implements algorithmic leveraging for linear regression using uniform and leverage score based on subsampling of rows
}
\examples{
x <- matrix(rnorm(500),100,5)
y <- x\%*\%rep(-1,5)+rnorm(100)
algo_leverage(x,y)

}
