% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elnet_coord.R
\name{elnet_coord}
\alias{elnet_coord}
\title{elnet_coord
ﬁts elastic net to data using coordinate descent algorithm.}
\usage{
elnet_coord(x, y, a, lambda, betahat = rep(0, NCOL(x)),
  maxiteration = 100)
}
\arguments{
\item{x}{n*p matrix;explanatory variables}

\item{y}{n*1 vector;response variable}

\item{a}{a constant;the l-2 regularization coefficient}

\item{lambda}{a constant;the l-1 and l-2 regularization coefficient}

\item{betahat}{p*1 vector;the initial regression coefficient;default is a zero vector}

\item{maxiteration}{maximum number of iteration allowed}
}
\value{
betahat the estimated regression coefficient
}
\description{
elnet_coord
ﬁts elastic net to data using coordinate descent algorithm.
}
