% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_ols.R
\name{solve_ols}
\alias{solve_ols}
\title{solve_ols
Solve linear system by Gauss-Seidel, Jacobi(sequential), or Jacobi(parallel) method}
\usage{
solve_ols(A, b, cores = 1, method, iteration)
}
\arguments{
\item{A}{n*p matrix of the input system Ax=b}

\item{b}{n*1 vector of the input system Ax=b}

\item{cores}{Number of cores used in parallel computing of Jacobi Mathod. If cores=1,use sequential method, otherwise use parallel computing.}

\item{method}{Method to use. "GS"=Gauss-Seidel,"Jacobi"=Jacobi}

\item{iteration}{Number of iterations}
}
\value{
xi n*1 vector, solution of the input system Ax=b
}
\description{
solve_ols
Solve linear system by Gauss-Seidel, Jacobi(sequential), or Jacobi(parallel) method
}
\examples{
a=matrix(rnorm(9),3,3)
v=rep(1,3)
b=a\%*\%v
solve_ols(a,b,method = "GS",iteration=100)
solve_ols(a,b,cores=1,method = "Jacobi",iteration=100)
solve_ols(a,b,cores=2,method = "Jacobi",iteration=100)
}
